+++ prover/crates/bin/prover_cli/src/commands/insert_batch.rs
@@ -31,7 +31,7 @@ pub async fn run(args: Args, config: ProverCLIConfig) -> anyhow::Result<()> {
 
     let protocol_version_patch = VersionPatch(args.patch);
 
-    conn.fri_witness_generator_dal()
+    conn.fri_basic_witness_generator_dal()
         .save_witness_inputs(
             args.number,
             &format!("witness_inputs_{}", args.number.0),
