+++ prover/crates/bin/prover_fri_gateway/src/proof_gen_data_fetcher.rs
@@ -49,7 +49,7 @@ impl ProofGenDataFetcher {
             .await;
 
         connection
-            .fri_witness_generator_dal()
+            .fri_basic_witness_generator_dal()
             .save_witness_inputs(data.l1_batch_number, &witness_inputs, data.protocol_version)
             .await;
     }
