+++ prover/crates/bin/witness_generator/src/rounds/leaf_aggregation/artifacts.rs
@@ -91,7 +91,7 @@ impl ArtifactsManager for LeafAggregation {
         let mut transaction = prover_connection.start_transaction().await.unwrap();
         let number_of_dependent_jobs = blob_urls.circuit_ids_and_urls.len();
         let protocol_version_id = transaction
-            .fri_witness_generator_dal()
+            .fri_basic_witness_generator_dal()
             .protocol_version_for_l1_batch(artifacts.block_number)
             .await;
         tracing::info!(
@@ -118,7 +118,7 @@ impl ArtifactsManager for LeafAggregation {
             artifacts.circuit_id,
         );
         transaction
-            .fri_witness_generator_dal()
+            .fri_node_witness_generator_dal()
             .update_node_aggregation_jobs_url(
                 artifacts.block_number,
                 get_recursive_layer_circuit_id_for_base_layer(artifacts.circuit_id),
@@ -134,7 +134,7 @@ impl ArtifactsManager for LeafAggregation {
             artifacts.circuit_id,
         );
         transaction
-            .fri_witness_generator_dal()
+            .fri_leaf_witness_generator_dal()
             .mark_leaf_aggregation_as_successful(job_id, started_at.elapsed())
             .await;
