+++ prover/crates/bin/witness_generator/src/rounds/leaf_aggregation/mod.rs
@@ -239,7 +239,7 @@ impl JobManager for LeafAggregation {
         let Some(metadata) = connection_pool
             .connection()
             .await?
-            .fri_witness_generator_dal()
+            .fri_leaf_witness_generator_dal()
             .get_next_leaf_aggregation_job(protocol_version, &pod_name)
             .await
         else {
