+++ prover/crates/bin/witness_generator/src/rounds/recursion_tip/artifacts.rs
@@ -108,7 +108,7 @@ impl ArtifactsManager for RecursionTip {
         let mut prover_connection = connection_pool.connection().await?;
         let mut transaction = prover_connection.start_transaction().await?;
         let protocol_version_id = transaction
-            .fri_witness_generator_dal()
+            .fri_basic_witness_generator_dal()
             .protocol_version_for_l1_batch(L1BatchNumber(job_id))
             .await;
         transaction
@@ -126,7 +126,7 @@ impl ArtifactsManager for RecursionTip {
             .await;
 
         transaction
-            .fri_witness_generator_dal()
+            .fri_recursion_tip_witness_generator_dal()
             .mark_recursion_tip_job_as_successful(L1BatchNumber(job_id), started_at.elapsed())
             .await;
