--- prover/crates/lib/prover_dal/.sqlx/query-340a84063719f9b837a61cbc63368ba43ba128a00a194dce65015bd9eeb1b42f.json
@@ -1,17 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                INSERT INTO\n                scheduler_witness_jobs_fri (\n                    l1_batch_number,\n                    scheduler_partial_input_blob_url,\n                    protocol_version,\n                    status,\n                    created_at,\n                    updated_at,\n                    protocol_version_patch\n                )\n                VALUES\n                ($1, $2, $3, 'waiting_for_proofs', NOW(), NOW(), $4)\n                ON CONFLICT (l1_batch_number) DO\n                UPDATE\n                SET\n                updated_at = NOW()\n                ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Int8",
-        "Text",
-        "Int4",
-        "Int4"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "340a84063719f9b837a61cbc63368ba43ba128a00a194dce65015bd9eeb1b42f"
-}
