--- prover/crates/lib/prover_dal/.sqlx/query-37fa629a87113f13c89ce5c1a8fffd15b636d553794900596528e8fb03ca5aed.json
@@ -1,17 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                INSERT INTO\n                recursion_tip_witness_jobs_fri (\n                    l1_batch_number,\n                    status,\n                    number_of_final_node_jobs,\n                    protocol_version,\n                    created_at,\n                    updated_at,\n                    protocol_version_patch\n                )\n                VALUES\n                ($1, 'waiting_for_proofs', $2, $3, NOW(), NOW(), $4)\n                ON CONFLICT (l1_batch_number) DO\n                UPDATE\n                SET\n                updated_at = NOW()\n                ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Int8",
-        "Int4",
-        "Int4",
-        "Int4"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "37fa629a87113f13c89ce5c1a8fffd15b636d553794900596528e8fb03ca5aed"
-}
