+++ prover/crates/lib/prover_dal/.sqlx/query-5fb1d2533749420889f9cf3b5ec326383ac71c031cf2d0a9752595761f5baa7f.json
@@ -1,0 +1,17 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            INSERT INTO\n            scheduler_witness_jobs_fri (\n                l1_batch_number,\n                scheduler_partial_input_blob_url,\n                protocol_version,\n                status,\n                created_at,\n                updated_at,\n                protocol_version_patch\n            )\n            VALUES\n            ($1, $2, $3, 'waiting_for_proofs', NOW(), NOW(), $4)\n            ON CONFLICT (l1_batch_number) DO\n            UPDATE\n            SET\n            updated_at = NOW()\n            ",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Text",
+        "Int4",
+        "Int4"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "5fb1d2533749420889f9cf3b5ec326383ac71c031cf2d0a9752595761f5baa7f"
+}
