+++ prover/crates/lib/prover_dal/.sqlx/query-d72c7beede026491202626a897e3aaee2b3708f40799a7aeab8e143b0b534c4f.json
@@ -1,0 +1,17 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            INSERT INTO\n            recursion_tip_witness_jobs_fri (\n                l1_batch_number,\n                status,\n                number_of_final_node_jobs,\n                protocol_version,\n                created_at,\n                updated_at,\n                protocol_version_patch\n            )\n            VALUES\n            ($1, 'waiting_for_proofs', $2, $3, NOW(), NOW(), $4)\n            ON CONFLICT (l1_batch_number) DO\n            UPDATE\n            SET\n            updated_at = NOW()\n            ",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Int4",
+        "Int4",
+        "Int4"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "d72c7beede026491202626a897e3aaee2b3708f40799a7aeab8e143b0b534c4f"
+}
