--- prover/crates/lib/prover_dal/.sqlx/query-f294a1c32ffb957c901dcdfa942ea8e04c2c28771aa50a9e3ebb95c5e428be0e.json
@@ -1,19 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                    INSERT INTO\n                    leaf_aggregation_witness_jobs_fri (\n                        l1_batch_number,\n                        circuit_id,\n                        closed_form_inputs_blob_url,\n                        number_of_basic_circuits,\n                        protocol_version,\n                        status,\n                        created_at,\n                        updated_at,\n                        protocol_version_patch\n                    )\n                    VALUES\n                    ($1, $2, $3, $4, $5, 'waiting_for_proofs', NOW(), NOW(), $6)\n                    ON CONFLICT (l1_batch_number, circuit_id) DO\n                    UPDATE\n                    SET\n                    updated_at = NOW()\n                    ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Int8",
-        "Int2",
-        "Text",
-        "Int4",
-        "Int4",
-        "Int4"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "f294a1c32ffb957c901dcdfa942ea8e04c2c28771aa50a9e3ebb95c5e428be0e"
-}
